# nestjs-backend/swagger-doc.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/asset/{ticker}": {
      "get": {
        "operationId": "AssetController_getByTicker",
        "summary": "Get one Asset by ticker",
        "parameters": [
          {
            "name": "ticker",
            "required": true,
            "in": "path",
            "description": "The ticker symbol of the asset",
            "example": "AAPL",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Asset details by ticker",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "_id": "3",
                    "ticker": "AAPL",
                    "name": "Apple",
                    "type": "STOCK",
                    "geography": "USA",
                    "industry": "Technology"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Asset Service"
        ]
      },
      "put": {
        "operationId": "AssetController_update",
        "summary": "Update an Asset by ticker",
        "parameters": [
          {
            "name": "ticker",
            "required": true,
            "in": "path",
            "description": "The ticker symbol of the asset to update",
            "example": "MSFT",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Updated asset details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetDto"
              },
              "examples": {
                "example1": {
                  "summary": "Updated Asset Example",
                  "value": {
                    "ticker": "MSFT",
                    "name": "Microsoft Corporation",
                    "type": "STOCK",
                    "geography": "USA",
                    "industry": "Technology"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Asset updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "_id": "1",
                    "ticker": "AAPL",
                    "name": "Apple",
                    "type": "STOCK",
                    "geography": "USA",
                    "industry": "Technology"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Asset Service"
        ]
      },
      "delete": {
        "operationId": "AssetController_delete",
        "summary": "Delete an Asset by ticker",
        "parameters": [
          {
            "name": "ticker",
            "required": true,
            "in": "path",
            "description": "The ticker symbol of the asset to delete",
            "example": "JPM",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": "Asset deleted successfully"
          }
        },
        "tags": [
          "Asset Service"
        ]
      }
    },
    "/assetprice": {
      "get": {
        "operationId": "AssetPriceController_getAll",
        "summary": "Get all AssetPrices",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all asset prices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "example": [
                      {
                        "_id": "60f6c72e0a9a2344a8d56fdb",
                        "ticker": "AMGN",
                        "company": "Amgen",
                        "sector": "Biopharmaceutical",
                        "todayClose": 245.67,
                        "yesterdayClose": 244.98,
                        "date": "2024-12-01T00:00:00.000Z"
                      },
                      {
                        "_id": "60f6c72e0a9a2344a8d56fdb",
                        "ticker": "DOW",
                        "company": "Dow Chemical",
                        "sector": "Chemical",
                        "todayClose": 56.45,
                        "yesterdayClose": 55.78,
                        "date": "2024-12-01T00:00:00.000Z"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "tags": [
          "AssetPrice Service"
        ]
      },
      "post": {
        "operationId": "AssetPriceController_create",
        "summary": "Create an AssetPrice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Create an asset price",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetPrice"
              },
              "examples": {
                "example1": {
                  "summary": "Create assetprice example",
                  "value": {
                    "ticker": "MSFT",
                    "name": "Microsoft Corporation",
                    "sector": "Information Technology",
                    "todayClose": 231.86000061035156,
                    "yesterdayClose": 228.8699951171875,
                    "date": "2024-12-01T00:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Asset price successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "_id": "60f6c72e0a9a2344a8d56fdb",
                    "ticker": "MSFT",
                    "company": "Microsoft Corporation",
                    "sector": "Information Technology",
                    "todayClose": 231.86000061035156,
                    "yesterdayClose": 228.8699951171875,
                    "date": "2024-12-01T00:00:00.000Z"
                  }
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetPrice"
                }
              }
            }
          }
        },
        "tags": [
          "AssetPrice Service"
        ]
      }
    },
    "/assetprice/{ticker}": {
      "get": {
        "operationId": "AssetPriceController_getByTicker",
        "summary": "Get an array of all AssetPrices with ticker",
        "parameters": [
          {
            "name": "ticker",
            "required": true,
            "in": "path",
            "description": "The ticker symbol of the asset",
            "example": "AAPL",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Asset prices for a specific ticker",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "example": [
                      {
                        "_id": "60f6c72e0a9a2344a8d56fdb",
                        "ticker": "AMGN",
                        "company": "Amgen",
                        "sector": "Biopharmaceutical",
                        "todayClose": 245.67,
                        "yesterdayClose": 244.98,
                        "date": "2024-12-01T00:00:00.000Z"
                      },
                      {
                        "_id": "638029d0a9a2344a8d56fdb",
                        "ticker": "AAPL",
                        "company": "Apple",
                        "sector": "Technology",
                        "todayClose": 130.25,
                        "yesterdayClose": 125.98,
                        "date": "2024-12-01T00:00:00.000Z"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "tags": [
          "AssetPrice Service"
        ]
      }
    },
    "/assetprice/{ticker}/{date}": {
      "get": {
        "operationId": "AssetPriceController_getByTickerAndDate",
        "summary": "Get an AssetPrice with ticker and date",
        "parameters": [
          {
            "name": "ticker",
            "required": true,
            "in": "path",
            "description": "The ticker symbol of the asset",
            "example": "AAPL",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "path",
            "description": "The price of asset at a specific date",
            "example": "2024-12-01T00:00:00.000Z",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Asset price for a specific ticker and date",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "_id": "60f6c72e0a9a2344a8d56fdb",
                    "ticker": "AMGN",
                    "company": "Amgen",
                    "sector": "Biopharmaceutical",
                    "todayClose": 245.67,
                    "yesterdayClose": 244.98,
                    "date": "2024-12-01T00:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "AssetPrice Service"
        ]
      },
      "put": {
        "operationId": "AssetPriceController_update",
        "summary": "Update an AssetPrice by ticker and date",
        "parameters": [
          {
            "name": "ticker",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetPriceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetPrice"
                }
              }
            }
          }
        },
        "tags": [
          "AssetPrice Service"
        ]
      },
      "delete": {
        "operationId": "AssetPriceController_delete",
        "summary": "Delete an AssetPrice by ticker and date",
        "parameters": [
          {
            "name": "ticker",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AssetPrice Service"
        ]
      }
    },
    "/assetprice/all/excluding/tickers": {
      "get": {
        "operationId": "AssetPriceController_getAllExcept",
        "summary": "Get all AssetPrice excluding specified",
        "parameters": [
          {
            "name": "exclusions",
            "required": false,
            "in": "query",
            "description": "Tickers of excluded assets in the query",
            "example": "[\"AAPL\", \"MSFT\"]",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of asset prices excluding specified tickers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "example": [
                      {
                        "_id": "60f6c72e0a9a2344a8d56fdb",
                        "ticker": "AMGN",
                        "company": "Amgen",
                        "sector": "Biopharmaceutical",
                        "todayClose": 245.67,
                        "yesterdayClose": 244.98,
                        "date": "2024-12-01T00:00:00.000Z"
                      },
                      {
                        "_id": "60f6c72e0a9a2344a8d56fdb",
                        "ticker": "DOW",
                        "company": "Dow Chemical",
                        "sector": "Chemical",
                        "todayClose": 56.45,
                        "yesterdayClose": 55.78,
                        "date": "2024-12-01T00:00:00.000Z"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "tags": [
          "AssetPrice Service"
        ]
      }
    },
    "/assetprice/all/from/tickers": {
      "get": {
        "operationId": "AssetPriceController_getFrom",
        "summary": "Get all AssetPrice inclusive only of specified",
        "parameters": [
          {
            "name": "inclusions",
            "required": false,
            "in": "query",
            "description": "Tickers of included assets in the query",
            "example": "[\"AAPL\", \"MSFT\"]",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of asset prices for specified tickers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "example": [
                      {
                        "_id": "60f6c72e0a9a2344a8d56fdb",
                        "ticker": "AMGN",
                        "company": "Amgen",
                        "sector": "Biopharmaceutical",
                        "todayClose": 245.67,
                        "yesterdayClose": 244.98,
                        "date": "2024-12-01T00:00:00.000Z"
                      },
                      {
                        "_id": "60f6c72e0a9a2344a8d56fdb",
                        "ticker": "DOW",
                        "company": "Dow Chemical",
                        "sector": "Chemical",
                        "todayClose": 56.45,
                        "yesterdayClose": 55.78,
                        "date": "2024-12-01T00:00:00.000Z"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "tags": [
          "AssetPrice Service"
        ]
      }
    },
...
  },
  "info": {
    "title": "Portfolio Optimizer Application API",
    "description": "core application in nestjs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "AssetDto": {
        "type": "object",
        "properties": {
          "ticker": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "enum": [
              "STOCK",
              "BOND"
            ],
            "type": "string"
          },
          "geography": {
            "type": "string"
          },
          "industry": {
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "name",
          "type",
          "geography",
          "industry"
        ]
      },
      "AssetPrice": {
        "type": "object",
        "properties": {}
      },
      "AssetPriceDto": {
        "type": "object",
        "properties": {
          "ticker": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "sector": {
            "type": "string"
          },
          "todayClose": {
            "type": "number"
          },
          "yesterdayClose": {
            "type": "number"
          },
          "date": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "company",
          "sector",
          "todayClose",
          "yesterdayClose",
          "date"
        ]
      },
...
    }
  }
}